Store draw commands on client so they can be replayed without bothering the server; surface 1 basically works, creating new displays fails (and should it?)

+++ src/vidlib/proxy/imagepng.c
@@ -1,6 +1,6 @@
 /*
  *  Crafted by Jim Buckeyne
- *   (c)1999-2006++ Freedom Collective
+ *	(c)1999-2006++ Freedom Collective
  * 
  * Load PNG images into an internal Image structure.
  * 
@@ -29,11 +29,11 @@
 
 typedef struct ImagePngRawData_tag
 {
-  // The buffer to "read" from
-  _8 **r_data;
-  // The buffer size
-  int *r_size;
-  int alloced;
+	// The buffer to "read" from
+	_8 **r_data;
+	// The buffer size
+	size_t *r_size;
+	size_t alloced;
 }ImagePngRawData;
 
 IMAGE_NAMESPACE
@@ -43,11 +43,11 @@ namespace loader {
 
 //--------------------------------------
 static void CPROC ImagePngWrite(png_structp png,
-                           png_bytep   data,
-                           png_size_t  length)
+									png_bytep	data,
+									png_size_t  length)
 {
 	// add this length, data into the buffer...
-   // reallcoate buffer by 4096 until complete.
+	// reallcoate buffer by 4096 until complete.
 	//sg_pStream->write(length, data);
 #if ( PNG_LIBPNG_VER > 10405 )
 	ImagePngRawData *self = (ImagePngRawData *)png_get_io_ptr( png );
@@ -69,15 +69,15 @@ static void CPROC ImagePngWrite(png_structp png,
 		}
 	}
 	MemCpy( (*self->r_data)+(*self->r_size), data, length );
-   (*self->r_size) = (*self->r_size) + length;
+	(*self->r_size) = (*self->r_size) + length;
 }
 
 
 //--------------------------------------
 static void CPROC ImagePngFlush(png_structp png_ptr)
 {
-   // not eally needed a flush here.
-   //
+	// not eally needed a flush here.
+	//
 }
 
 
@@ -87,84 +87,84 @@ static void CPROC NotSoFatalError( png_structp png_ptr, png_const_charp c )
 	lprintf( WIDE("Error in PNG stuff: %s"), c );
 }
 
-LOGICAL PngImageFile ( Image pImage, _8 ** buf, int *size)
+LOGICAL PngImageFile ( Image pImage, _8 ** buf, size_t *size)
 {
-   png_structp png_ptr;
-   png_infop info_ptr;
-   //png_infop end_info;
+	png_structp png_ptr;
+	png_infop info_ptr;
+	//png_infop end_info;
 	ImagePngRawData raw;
 
 	if( !pImage->width || !pImage->height )
-      return FALSE;
+		return FALSE;
 
-   png_ptr = png_create_write_struct( PNG_LIBPNG_VER_STRING, NULL,
+	png_ptr = png_create_write_struct( PNG_LIBPNG_VER_STRING, NULL,
 				  NULL, NULL);
-   if (!png_ptr)
+	if (!png_ptr)
 		return FALSE;
 	png_set_error_fn( png_ptr, (void*)NotSoFatalError //png_get_error_ptr( png_ptr )
 						 , NotSoFatalError, NotSoFatalError );
 	//png_ptr->error_fn = NotSoFatalError;
-   info_ptr = png_create_info_struct(png_ptr);
-   if (!info_ptr)
-   {
+	info_ptr = png_create_info_struct(png_ptr);
+	if (!info_ptr)
+	{
 no_mem2:
-      png_destroy_write_struct(&png_ptr,
-         (png_infopp)NULL);
-      return FALSE;
-   }
+		png_destroy_write_struct(&png_ptr,
+			(png_infopp)NULL);
+		return FALSE;
+	}
 
-   //end_info = png_create_info_struct(png_ptr);
-   //if (!end_info)
-   //{
-   //   png_destroy_write_struct(&png_ptr, &info_ptr,
-   //     (png_infopp)NULL);
-   //   return NULL;
-   //}
-   {
-      raw.r_data = buf;
+	//end_info = png_create_info_struct(png_ptr);
+	//if (!end_info)
+	//{
+	//	png_destroy_write_struct(&png_ptr, &info_ptr,
+	//	  (png_infopp)NULL);
+	//	return NULL;
+	//}
+	{
+		raw.r_data = buf;
 		raw.r_size = size;
-      raw.alloced = 0;
-      png_set_write_fn( png_ptr, &raw, ImagePngWrite, ImagePngFlush );
-   }
+		raw.alloced = 0;
+		png_set_write_fn( png_ptr, &raw, ImagePngWrite, ImagePngFlush );
+	}
 
-   // Set the compression level, image filters, and compression strategy...
-   //png_ptr->flags        |= PNG_FLAG_ZLIB_CUSTOM_STRATEGY;
-   //png_ptr->zlib_strategy = Z_DEFAULT_STRATEGY;
-   png_set_compression_window_bits(png_ptr, 15);
-   png_set_compression_level(png_ptr, /*0-9*/ 3 ); // gzip level?
-   png_set_filter(png_ptr, 0, PNG_FILTER_NONE);
+	// Set the compression level, image filters, and compression strategy...
+	//png_ptr->flags		  |= PNG_FLAG_ZLIB_CUSTOM_STRATEGY;
+	//png_ptr->zlib_strategy = Z_DEFAULT_STRATEGY;
+	png_set_compression_window_bits(png_ptr, 15);
+	png_set_compression_level(png_ptr, /*0-9*/ 3 ); // gzip level?
+	png_set_filter(png_ptr, 0, PNG_FILTER_NONE);
 
-   png_set_bgr( png_ptr );
-   //png_set_swap_alpha( png_ptr );
+	png_set_bgr( png_ptr );
+	//png_set_swap_alpha( png_ptr );
 	png_set_IHDR(png_ptr, info_ptr,
-					 pImage->width, pImage->height,               // the width & height
+					 pImage->width, pImage->height,					// the width & height
 					 8, PNG_COLOR_TYPE_RGB_ALPHA, // bit_depth, color_type,
-					 PNG_INTERLACE_NONE,          // no interlace
-					 PNG_COMPRESSION_TYPE_BASE,   // compression type
-					 PNG_FILTER_TYPE_BASE);       // filter type
-   png_write_info(png_ptr, info_ptr);
+					 PNG_INTERLACE_NONE,			 // no interlace
+					 PNG_COMPRESSION_TYPE_BASE,	// compression type
+					 PNG_FILTER_TYPE_BASE);		 // filter type
+	png_write_info(png_ptr, info_ptr);
 
 	{
-      int row;
+		int row;
 		png_bytep * const row_pointers = (png_bytep*const)Allocate( sizeof( png_bytep ) * pImage->height );
 		for (row=0; row< pImage->height; row++)
 		{
 #ifdef _INVERT_IMAGE
 			row_pointers[row] = (png_bytep)(pImage->image + (pImage->height-row-1) * pImage->pwidth);
 #else
-		   row_pointers[row] = (png_bytep)(pImage->image + row * pImage->pwidth);
+			row_pointers[row] = (png_bytep)(pImage->image + row * pImage->pwidth);
 #endif
 		}
 		png_write_image(png_ptr, row_pointers);
-      Release( row_pointers );
+		Release( row_pointers );
 	}
 
-   png_write_end(png_ptr, info_ptr);
+	png_write_end(png_ptr, info_ptr);
 
-   png_destroy_write_struct( &png_ptr, &info_ptr
-                            );
+	png_destroy_write_struct( &png_ptr, &info_ptr
+									 );
 
-   return TRUE;
+	return TRUE;
 }
 #ifdef __cplusplus
 }//namespace loader {
